Revert "made new files for ephemeral.js store and the constant sources,  templates inside of it"
This reverts commit a6bf14d852ca45315350f2ee0a11bf91b85e770c.
