--- frontend/www/js/omegaup/grader/eventBus.js
@@ -1,2 +1,0 @@
-import Vue from 'vue';
-export default new Vue();
